New APi - some cleanup of demo
